Refactor: swap places folder button, file CallControls
