+++ package.json
@@ -9,6 +9,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
+    "switchbox": "^1.0.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
